+++ app/authorization/client/index.js
@@ -1,5 +1,6 @@
 import { hasAllPermission, hasAtLeastOnePermission, hasPermission, userHasAllPermission } from './hasPermission';
 import { hasRole } from './hasRole';
+import { AuthorizationUtils } from '../lib';
 import './usersNameChanged';
 import './requiresPermission.html';
 import './route';
@@ -12,4 +13,5 @@ export {
 	hasRole,
 	hasPermission,
 	userHasAllPermission,
+	AuthorizationUtils,
 };
