+++ app/authorization/client/views/index.js
@@ -1,3 +1,5 @@
+import './permissionsTable.html';
+import './permissionsTable';
 import './permissions.html';
 import './permissions';
 import './permissionsRole.html';
