+++ app/authorization/client/views/permissions.html
@@ -1,45 +1,3 @@
-<template name="permissionsTable">
-		<div class="content">
-			<table class="permission-grid secondary-background-color">
-
-				<thead class="content-background-color">
-					<tr>
-						<th class=" role-name permission-name border-component-color"></th>
-						{{#each role in allRoles}}
-						<th class="edit-icon role-name border-component-color" title="{{role.description}}">
-							<a class="permission-edit" href="{{pathFor "admin-permissions-edit" name=role._id}}">
-								<p class="permission-label">
-									<div class = "rotator">
-										{{role._id}}
-									</div>
-								</p>
-								<p class="permission-icon">
-									<i class="icon-edit"></i>
-								</p>
-							</a>
-						</th>
-						{{/each}}
-					</tr>
-				</thead>
-				<tbody>
-					{{#each permission in permissions}}
-					<tr class="admin-table-row">
-						<td class="permission-name border-component-color" title="{{permissionDescription permission}}">{{permissionName permission}}<br><span class = "id-styler">[ID: {{permission._id}}]</span></td>
-						{{#each role in allRoles}}
-						<td class="permission-checkbox border-component-color">
-							<input type="checkbox" name="perm[{{role._id}}][{{permission._id}}]" class="role-permission" value="1" checked="{{granted permission.roles role}}" data-role="{{role._id}}" data-permission="{{permission._id}}">
-						</td>
-						{{else}}
-						<tr class="table-no-click">
-							<td>{{_ "No_results_found_for"}} {{filter}}</td>
-						</tr>
-						{{/each}}
-					</tr>
-					{{/each}}
-				</tbody>
-			</table>
-		</div>
-</template>
 <template name="permissions">
 	<div class="page-container permissions-manager">
 		<section class="permissions-manager page-container page-home page-static page-list">
