+++ app/authorization/client/views/permissions.js
@@ -1,6 +1,5 @@
 import _ from 'underscore';
 import s from 'underscore.string';
-import { Meteor } from 'meteor/meteor';
 import { ReactiveDict } from 'meteor/reactive-dict';
 import { Tracker } from 'meteor/tracker';
 import { Template } from 'meteor/templating';
@@ -154,44 +153,6 @@ Template.permissions.onCreated(function() {
 		this.state.get('filter');
 		this.state.set('size', 50);
 	});
-});
-
-Template.permissionsTable.helpers({
-	granted(roles, role) {
-		return (roles && ~roles.indexOf(role._id) && 'checked') || null;
-	},
-
-	permissionName(permission) {
-		if (permission.level === CONSTANTS.SETTINGS_LEVEL) {
-			let path = '';
-			if (permission.group) {
-				path = `${ t(permission.group) } > `;
-			}
-			if (permission.section) {
-				path = `${ path }${ t(permission.section) } > `;
-			}
-			return `${ path }${ t(permission.settingId) }`;
-		}
-
-		return t(permission._id);
-	},
-
-	permissionDescription(permission) {
-		return t(`${ permission._id }_description`);
-	},
-});
-
-Template.permissionsTable.events({
-	'click .role-permission'(e) {
-		const permissionId = e.currentTarget.getAttribute('data-permission');
-		const role = e.currentTarget.getAttribute('data-role');
-
-		const permission = permissionId && ChatPermissions.findOne(permissionId);
-
-		const action = ~permission.roles.indexOf(role) ? 'authorization:removeRoleFromPermission' : 'authorization:addPermissionToRole';
-
-		return Meteor.call(action, permissionId, role);
-	},
 });
 
 Template.permissions.onRendered(() => {
