+++ ee/app/authorization/index.js
@@ -1,0 +1,1 @@
+import './lib';
