+++ ee/client/index.js
@@ -1,4 +1,5 @@
 import '../app/auditing/client/index';
+import '../app/authorization/index';
 import '../app/canned-responses/client/index';
 import '../app/engagement-dashboard/client/index';
 import '../app/license/client/index';
