+++ ee/server/index.js
@@ -1,6 +1,7 @@
 import '../app/models';
 import '../app/api-enterprise/server/index';
 import '../app/auditing/server/index';
+import '../app/authorization/index';
 import '../app/canned-responses/server/index';
 import '../app/engagement-dashboard/server/index';
 import '../app/ldap-enterprise/server/index';
